Set up the typescript scaffolding for the browser conversion tests
