+++ .gitignore
@@ -13,6 +13,7 @@ rpc/clients/typescript/ganache.log
 *.wasm
 browser/lib
 browser/ts/generated
+browser/dist/bundle.js
 examples/browser/dist/bundle.js
 integration-tests/browser/dist/bundle.js
 !integration-tests/data/standalone-0/keys/privKey
