+++ browser/.prettierrc
@@ -1,0 +1,6 @@
+{
+    "printWidth": 120,
+    "tabWidth": 4,
+    "singleQuote": true,
+    "trailingComma": "all"
+}
