+++ browser/package.json
@@ -5,7 +5,9 @@
     "main": "./lib/index.js",
     "license": "Apache-2.0",
     "scripts": {
-        "build": "yarn build:go && yarn build:generate && yarn build:ts",
+        "prettier": "prettier --write '**/*.{ts,tsx,json,md}' --config .prettierrc",
+        "build": "yarn build:go && yarn build:generate && yarn build:ts && yarn build:bundle",
+        "build:bundle": "node --max_old_space_size=3072 ./node_modules/webpack/bin/webpack.js --mode=development",
         "build:ts": "node --max_old_space_size=4096 ./node_modules/.bin/tsc -b",
         "build:generate": "go run ./scripts/generate_wasm_buffer.go",
         "build:go": "GOOS=js GOARCH=wasm go build -o ./wasm/main.wasm ./go/main.go",
@@ -19,8 +21,12 @@
         "@0x/ts-doc-gen": "^0.0.16",
         "@0x/tslint-config": "^4.0.0",
         "@types/base64-arraybuffer": "^0.1.0",
+        "prettier": "^1.19.1",
+        "ts-loader": "^6.2.1",
         "tslint": "5.11.0",
-        "typescript": "^3.5.3"
+        "typescript": "^3.5.3",
+        "webpack": "^4.41.5",
+        "webpack-cli": "^3.3.10"
     },
     "dependencies": {
         "@0x/order-utils": "^10.0.1",
