+++ browser/ts/browser_test.ts
@@ -1,0 +1,12 @@
+import {
+    configToWrapperConfig,
+    orderEventsHandlerToWrapperOrderEventsHandler,
+    signedOrderToWrapperSignedOrder,
+    wrapperAcceptedOrderInfoToAcceptedOrderInfo,
+    wrapperContractEventsToContractEvents,
+    wrapperOrderEventToOrderEvent,
+    wrapperRejectedOrderInfoToRejectedOrderInfo,
+    wrapperSignedOrderToSignedOrder,
+    wrapperValidationResultsToValidationResults,
+} from './encoding';
+import './wasm_exec';
